KULRICE-2343: deleting attachments doesn't work (was: Notes and attachments deletion doesn't work on maintenance documents) https://test.kuali.org/jira/browse/KULRICE-2343
The method for deleting AttachmentS was assuming that the note instance
variable would be set, and in fact it needs to be set in order to find the
attachment file and delete it.  Made sure the reference was set even if
the Attachment and Note hadn't been persisted yet.

AttachmentServiceImpl.java:
- in deleteAttachmentContents(...), throw a RuntimeException with a better 
  message if Attachment.note isn't set. 

Note.java:
- In addAttachment(...), set the Attachment.note instance to this

KualiDocumentActionBase.java:
- in deleteBONote(), don't refresh the Attachment's reference objects (the Note)
  unless the Note has already been persisted

+++ impl/src/main/java/org/kuali/rice/kns/service/impl/AttachmentServiceImpl.java
@@ -157,6 +157,7 @@ public class AttachmentServiceImpl implements AttachmentService {
     }
 
     public void deleteAttachmentContents(Attachment attachment) {
+    	if (attachment.getNote() == null) throw new RuntimeException("Attachment.note must be set in order to delete the attachment");
         String fullPathUniqueFileName = getDocumentDirectory(attachment.getNote().getRemoteObjectIdentifier()) + File.separator + attachment.getAttachmentIdentifier();
         File attachmentFile = new File(fullPathUniqueFileName);
         attachmentFile.delete();
